+++ astropy/time/core.py
@@ -9,6 +9,7 @@ astronomy.
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
+import itertools
 import fnmatch
 import time
 
@@ -660,6 +661,21 @@ class Time(object):
         copy of the JD arrays.
         """
         return self.copy()
+
+    def __iter__(self):
+        if self.isscalar:
+            raise TypeError('scalar {0!r} object is not iterable.'.format(
+                self.__class__.__name__))
+
+        def time_iter():
+            try:
+                for idx in itertools.count():
+                    yield self[idx]
+            except IndexError:
+                # Results in StopIteration
+                pass
+
+        return time_iter()
 
     def __getitem__(self, item):
         if self.isscalar: